Backend/requirements.txt
fastapi
uvicorn
pillow
exifread
numpy<2.0
opencv-python
imagehash
matplotlib
#torch
#torchvision
scikit-learn
python-multipart

piexif
hachoir
mediapipe==0.10.10
protobuf==3.20.3
#tensorflow-intel==2.15.0
tensorflow==2.15.0
deepface
scikit-image
ultralytics